backend/requirements.txt
fastapi
uvicorn
pytest
transformers
torch
llama-cpp-python
more-itertools